# openapi.yaml
openapi: 3.0.1
info:
  title: Paperless Rest Server
  version: '1.0'
paths:
  /api/saved_views:
    get:
      tags:
        - Config
      operationId: GetSavedViews
      parameters:
        - name: page
          in: query
          style: form
          schema:
            type: integer
            format: int32
        - name: page_size
          in: query
          style: form
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  next:
                    type: integer
                  previous:
                    type: integer
                  all:
                    type: array
                    items:
                      type: integer
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        name:
                          type: string
                        show_on_dashboard:
                          type: boolean
                        show_in_sidebar:
                          type: boolean
                        sort_field:
                          type: string
                        sort_reverse:
                          type: boolean
                        filter_rules:
                          type: array
                          items:
                            type: object
                            properties:
                              rule_type:
                                type: integer
                              value:
                                type: string
                            required:
                              - rule_type
                              - value
                        owner:
                          type: object
                          properties:
                            id:
                              type: integer
                            password:
                              type: string
                            last_login:
                              type: string
                            is_superuser:
                              type: boolean
                            username:
                              type: string
                            first_name:
                              type: string
                            last_name:
                              type: string
                            email:
                              type: string
                            is_staff:
                              type: boolean
                            is_active:
                              type: boolean
                            date_joined:
                              type: string
                            groups:
                              type: array
                              items: {}
                            user_permissions:
                              type: array
                              items:
                                type: integer
                          required:
                            - id
                            - password
                            - last_login
                            - is_superuser
                            - username
                            - first_name
                            - last_name
                            - email
                            - is_staff
                            - is_active
                            - date_joined
                            - groups
                            - user_permissions
                        user_can_change:
                          type: boolean
                      required:
                        - id
                        - name
                        - show_on_dashboard
                        - show_in_sidebar
                        - sort_field
                        - sort_reverse
                        - filter_rules
                        - owner
                        - user_can_change
                required:
                  - count
                  - next
                  - previous
                  - all
                  - results
              examples:
                default:
                  value: >-
                    {"count":6,"next":null,"previous":null,"all":[8,17,7,4,11,15],"results":[{"id":8,"name":"Correspondent
                    2","show_on_dashboard":false,"show_in_sidebar":false,"sort_field":"created","sort_reverse":true,"filter_rules":[{"rule_type":3,"value":"2"}],"owner":{"id":2,"password":"pbkdf2_sha256$390000$ElHGEQzOIS3Gv2wu8YDbSt$bWASwmUfHtFMZ6+rG+4rrrtejjC9RiwXm6xMi0bxMZo=","last_login":"2023-04-26T09:08:14Z","is_superuser":false,"username":"testuser","first_name":"","last_name":"","email":"test@user.com","is_staff":true,"is_active":true,"date_joined":"2021-01-03T07:37:04Z","groups":[],"user_permissions":[57,58,59,60,5,6,7,8,1,2,3,4,9,10,11,12,61,62,63,64,65,66,67,68,13,14,15,16,117,118,119,120,121,122,123,124,113,114,115,116,77,78,79,80,85,86,87,88,69,70,71,72,81,82,83,84,73,74,75,76,33,34,35,36,21,22,23,24,37,38,39,40,89,90,91,92,29,30,31,32,109,137,110,138,111,139,112,140,105,106,107,108,41,42,43,44,45,46,47,48,93,94,95,96,25,26,27,28,101,102,103,104,97,98,99,100,125,126,127,128,129,130,131,132,49,50,51,52,53,54,55,56,133,134,135,136,17,18,19,20]},"user_can_change":true},{"id":17,"name":"In
                    the Last
                    Month","show_on_dashboard":false,"show_in_sidebar":false,"sort_field":"created","sort_reverse":true,"filter_rules":[{"rule_type":20,"value":"created:[-1
                    month to
                    now]"}],"owner":{"id":2,"password":"pbkdf2_sha256$390000$ElHGEQzOIS3Gv2wu8YDbSt$bWASwmUfHtFMZ6+rG+4rrrtejjC9RiwXm6xMi0bxMZo=","last_login":"2023-04-26T09:08:14Z","is_superuser":false,"username":"testuser","first_name":"","last_name":"","email":"test@user.com","is_staff":true,"is_active":true,"date_joined":"2021-01-03T07:37:04Z","groups":[],"user_permissions":[57,58,59,60,5,6,7,8,1,2,3,4,9,10,11,12,61,62,63,64,65,66,67,68,13,14,15,16,117,118,119,120,121,122,123,124,113,114,115,116,77,78,79,80,85,86,87,88,69,70,71,72,81,82,83,84,73,74,75,76,33,34,35,36,21,22,23,24,37,38,39,40,89,90,91,92,29,30,31,32,109,137,110,138,111,139,112,140,105,106,107,108,41,42,43,44,45,46,47,48,93,94,95,96,25,26,27,28,101,102,103,104,97,98,99,100,125,126,127,128,129,130,131,132,49,50,51,52,53,54,55,56,133,134,135,136,17,18,19,20]},"user_can_change":true},{"id":7,"name":"Inbox","show_on_dashboard":true,"show_in_sidebar":true,"sort_field":"created","sort_reverse":true,"filter_rules":[{"rule_type":6,"value":"9"}],"owner":{"id":2,"password":"pbkdf2_sha256$390000$ElHGEQzOIS3Gv2wu8YDbSt$bWASwmUfHtFMZ6+rG+4rrrtejjC9RiwXm6xMi0bxMZo=","last_login":"2023-04-26T09:08:14Z","is_superuser":false,"username":"testuser","first_name":"","last_name":"","email":"test@user.com","is_staff":true,"is_active":true,"date_joined":"2021-01-03T07:37:04Z","groups":[],"user_permissions":[57,58,59,60,5,6,7,8,1,2,3,4,9,10,11,12,61,62,63,64,65,66,67,68,13,14,15,16,117,118,119,120,121,122,123,124,113,114,115,116,77,78,79,80,85,86,87,88,69,70,71,72,81,82,83,84,73,74,75,76,33,34,35,36,21,22,23,24,37,38,39,40,89,90,91,92,29,30,31,32,109,137,110,138,111,139,112,140,105,106,107,108,41,42,43,44,45,46,47,48,93,94,95,96,25,26,27,28,101,102,103,104,97,98,99,100,125,126,127,128,129,130,131,132,49,50,51,52,53,54,55,56,133,134,135,136,17,18,19,20]},"user_can_change":true},{"id":4,"name":"Recently
                    Added","show_on_dashboard":true,"show_in_sidebar":true,"sort_field":"created","sort_reverse":true,"filter_rules":[],"owner":{"id":2,"password":"pbkdf2_sha256$390000$ElHGEQzOIS3Gv2wu8YDbSt$bWASwmUfHtFMZ6+rG+4rrrtejjC9RiwXm6xMi0bxMZo=","last_login":"2023-04-26T09:08:14Z","is_superuser":false,"username":"testuser","first_name":"","last_name":"","email":"test@user.com","is_staff":true,"is_active":true,"date_joined":"2021-01-03T07:37:04Z","groups":[],"user_permissions":[57,58,59,60,5,6,7,8,1,2,3,4,9,10,11,12,61,62,63,64,65,66,67,68,13,14,15,16,117,118,119,120,121,122,123,124,113,114,115,116,77,78,79,80,85,86,87,88,69,70,71,72,81,82,83,84,73,74,75,76,33,34,35,36,21,22,23,24,37,38,39,40,89,90,91,92,29,30,31,32,109,137,110,138,111,139,112,140,105,106,107,108,41,42,43,44,45,46,47,48,93,94,95,96,25,26,27,28,101,102,103,104,97,98,99,100,125,126,127,128,129,130,131,132,49,50,51,52,53,54,55,56,133,134,135,136,17,18,19,20]},"user_can_change":true},{"id":11,"name":"Tag:
                    Another Sample
                    Tag","show_on_dashboard":false,"show_in_sidebar":true,"sort_field":"created","sort_reverse":true,"filter_rules":[{"rule_type":6,"value":"4"}],"owner":{"id":2,"password":"pbkdf2_sha256$390000$ElHGEQzOIS3Gv2wu8YDbSt$bWASwmUfHtFMZ6+rG+4rrrtejjC9RiwXm6xMi0bxMZo=","last_login":"2023-04-26T09:08:14Z","is_superuser":false,"username":"testuser","first_name":"","last_name":"","email":"test@user.com","is_staff":true,"is_active":true,"date_joined":"2021-01-03T07:37:04Z","groups":[],"user_permissions":[57,58,59,60,5,6,7,8,1,2,3,4,9,10,11,12,61,62,63,64,65,66,67,68,13,14,15,16,117,118,119,120,121,122,123,124,113,114,115,116,77,78,79,80,85,86,87,88,69,70,71,72,81,82,83,84,73,74,75,76,33,34,35,36,21,22,23,24,37,38,39,40,89,90,91,92,29,30,31,32,109,137,110,138,111,139,112,140,105,106,107,108,41,42,43,44,45,46,47,48,93,94,95,96,25,26,27,28,101,102,103,104,97,98,99,100,125,126,127,128,129,130,131,132,49,50,51,52,53,54,55,56,133,134,135,136,17,18,19,20]},"user_can_change":true},{"id":15,"name":"View
                    ASN not
                    empty","show_on_dashboard":false,"show_in_sidebar":false,"sort_field":"created","sort_reverse":true,"filter_rules":[{"rule_type":18,"value":"false"}],"owner":{"id":2,"password":"pbkdf2_sha256$390000$ElHGEQzOIS3Gv2wu8YDbSt$bWASwmUfHtFMZ6+rG+4rrrtejjC9RiwXm6xMi0bxMZo=","last_login":"2023-04-26T09:08:14Z","is_superuser":false,"username":"testuser","first_name":"","last_name":"","email":"test@user.com","is_staff":true,"is_active":true,"date_joined":"2021-01-03T07:37:04Z","groups":[],"user_permissions":[57,58,59,60,5,6,7,8,1,2,3,4,9,10,11,12,61,62,63,64,65,66,67,68,13,14,15,16,117,118,119,120,121,122,123,124,113,114,115,116,77,78,79,80,85,86,87,88,69,70,71,72,81,82,83,84,73,74,75,76,33,34,35,36,21,22,23,24,37,38,39,40,89,90,91,92,29,30,31,32,109,137,110,138,111,139,112,140,105,106,107,108,41,42,43,44,45,46,47,48,93,94,95,96,25,26,27,28,101,102,103,104,97,98,99,100,125,126,127,128,129,130,131,132,49,50,51,52,53,54,55,56,133,134,135,136,17,18,19,20]},"user_can_change":true}]}
    post:
      tags:
        - Config
      operationId: CreateSavedViews
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                show_on_dashboard:
                  type: boolean
                show_in_sidebar:
                  type: boolean
                filter_rules:
                  type: array
                  items:
                    type: object
                    properties:
                      rule_type:
                        type: integer
                      value:
                        type: string
                    required:
                      - rule_type
                      - value
                sort_reverse:
                  type: boolean
                sort_field:
                  type: string
              required:
                - name
                - show_on_dashboard
                - show_in_sidebar
                - filter_rules
                - sort_reverse
                - sort_field
            examples:
              default:
                value: >-
                  {"name":"test","show_on_dashboard":true,"show_in_sidebar":true,"filter_rules":[{"rule_type":19,"value":"test"},{"rule_type":28,"value":"1"},{"rule_type":30,"value":"1"}],"sort_reverse":true,"sort_field":"created"}
      responses:
        '200':
          description: Success
...